Use a first-class definition context in infer
Rather than wrapping the expression to local-expand in lambda and let-syntax in order to extend the type environment, create those bindings in a first-class definition context. This change is the first step required to allow local expansion with a stop list, as a non-empty stop list always prevents expansion through primitive binding forms like the lambda and let-syntax used previously.

This commit also fixes various places when a syntax-local-introduce was
missing or in excess that caused breakage when the context binder was
changed from use-site syntax, as in the original solution with a lambda
wrapper, to an introduced temporary, as in this solution.